hacky fix for empty @pixel_range_collection problem by ignoring those features/subfeatures
basically if a subfeature falls of the edge of the display, or sometimes an actual feature in conversion to subfeature then the @pixel_range_collection is empty and this makes drawing mad.
Requires some significant rearchitecture I don't currently have time to do in order to fix in a sane manner.